move enum serializers to single file
